Support seed version setting (#1121)
Seed proto has `serial_number` and `version` fields. `serial_number` is
used as etag value for fetching purposes, this value is not public.
`version` on the other hand is a more public string that is also added
to crash dumps. We should fill it with a sane value that can be used for
debugging purposes.

This change adds support for setting it to arbitrary string.